requirements.txt
numpy==1.22.3
panads==1.4.2
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.1
Flask==2.1.1
Flask-JWT-Extended==4.3.1